Fix intermittent python crash when closing device
When close is called on a HID device and the current read is aborted,
the call to ReadFile does not get canceled.  This is a problem since
once the InputReportReaderThread's run method returns, the buffers
passed to ReadFile get released, even though they are still in
use.  If another packet arrives when these buffers are freed it causes
python to crash.  This patch makes sure that ReadFile is properly
canceled when close() is called and it aborts the current read.