Fix assert on HID device disconnect
If the current HID device is removed pywinusb will assert and exit
without properly cleaning up resources.  This patch fixes that problem
by removing the assert, along with the call to hid_object.close(), since
if called, it would trigger another error from trying to join itself.

This patch also makes the call to abort on the reader thread more
robust.  It also adds stronger guarantees to resource cleanup
in the InputReportReaderThread.