Allow exclusive device access in HidDevice
Add an option to HidDevice.open() to open with exclusive access.
By default devices will be opened in shared mode.  If the shared
parameter is set to False then the HID device will be opened
with exclusive access.

Note:  Depending on which programs are running, some HID devices will
already be opened in shared mode.  Trying to open one of these devices
with exclusive access will result in failure being returned by
CreateFile.