Fix intermittent lockup on close
When aborting execution of the input reader thread set "__abort" to
true, even if __h_read_event is null. This prevents a race condition
which occurs if abort is called before the input reader thread has
created its read event. If this race condition occurs then __abort
will not get set to true and the input reader thread will not
terminate. This causes close() to get stuck when trying to
join this thread.